Generate list of autosuggestions
Parent's state gets updated with AJAX autosuggestion data based on text input value.
AJAX autosuggestion data gets passed down as prop.
Use a conditional to wait for fetch.
Slice autosuggestion array into a length of 5.
Use a conditional to return empty div if user hasn't typed anything.
Generate a component for each value of the autosuggestion array.
Define a key for generated list items.
Pass current value as a prop to the list item.
In the list item component, display the current value and append the current value to the Google url.
Add proptypes to components.